@@ -1,9 +1,5 @@
 module golang.org/x/mod
 
-go 1.12
+go 1.17
 
-require (
-	golang.org/x/crypto v0.0.0-20191011191535-87dc89f01550
-	golang.org/x/tools v0.0.0-20191119224855-298f0cb1881e
-	golang.org/x/xerrors v0.0.0-20191011141410-1b5146add898
-)
+require golang.org/x/tools v0.1.12 // tagx:ignore
