@@ -78,8 +78,8 @@ var checkPathTests = []struct {
 	{"x.y/\xFFz", false, false, false},
 	{"/x.y/z", false, false, false},
 	{"x./z", false, false, false},
-	{".x/z", false, false, true},
-	{"-x/z", false, false, false},
+	{".x/z", false, true, true},
+	{"-x/z", false, false, true},
 	{"x..y/z", true, true, true},
 	{"x.y/z/../../w", false, false, false},
 	{"x.y//z", false, false, false},
@@ -104,7 +104,7 @@ var checkPathTests = []struct {
 	{"x.y(/z", false, false, true},
 	{"x.y)/z", false, false, true},
 	{"x.y*/z", false, false, false},
-	{"x.y+/z", false, false, true},
+	{"x.y+/z", false, true, true},
 	{"x.y,/z", false, false, true},
 	{"x.y-/z", true, true, true},
 	{"x.y./zt", false, false, false},
@@ -134,7 +134,7 @@ var checkPathTests = []struct {
 	{"x.y/z(", false, false, true},
 	{"x.y/z)", false, false, true},
 	{"x.y/z*", false, false, false},
-	{"x.y/z+", false, false, true},
+	{"x.y/z++", false, true, true},
 	{"x.y/z,", false, false, true},
 	{"x.y/z-", true, true, true},
 	{"x.y/z.t", true, true, true},
@@ -184,8 +184,8 @@ var checkPathTests = []struct {
 	{"./y", false, false, false},
 	{"x:y", false, false, false},
 	{`\temp\foo`, false, false, false},
-	{".gitignore", false, false, true},
-	{".github/ISSUE_TEMPLATE", false, false, true},
+	{".gitignore", false, true, true},
+	{".github/ISSUE_TEMPLATE", false, true, true},
 	{"x☺y", false, false, false},
 }
 
@@ -353,6 +353,8 @@ func TestMatchPrefixPatterns(t *testing.T) {
 		globs, target string
 		want          bool
 	}{
+		{"", "rsc.io/quote", false},
+		{"/", "rsc.io/quote", false},
 		{"*/quote", "rsc.io/quote", true},
 		{"*/quo", "rsc.io/quote", false},
 		{"*/quo??", "rsc.io/quote", true},
@@ -360,17 +362,21 @@ func TestMatchPrefixPatterns(t *testing.T) {
 		{"*quo*", "rsc.io/quote", false},
 		{"rsc.io", "rsc.io/quote", true},
 		{"*.io", "rsc.io/quote", true},
-		{"rsc.io/", "rsc.io/quote", false},
+		{"rsc.io/", "rsc.io/quote", true},
 		{"rsc", "rsc.io/quote", false},
 		{"rsc*", "rsc.io/quote", true},
 
 		{"rsc.io", "rsc.io/quote/v3", true},
 		{"*/quote", "rsc.io/quote/v3", true},
+		{"*/quote/", "rsc.io/quote/v3", true},
 		{"*/quote/*", "rsc.io/quote/v3", true},
+		{"*/quote/*/", "rsc.io/quote/v3", true},
 		{"*/v3", "rsc.io/quote/v3", false},
 		{"*/*/v3", "rsc.io/quote/v3", true},
 		{"*/*/*", "rsc.io/quote/v3", true},
+		{"*/*/*/", "rsc.io/quote/v3", true},
 		{"*/*/*", "rsc.io/quote", false},
+		{"*/*/*/", "rsc.io/quote", false},
 
 		{"*/*/*,,", "rsc.io/quote", false},
 		{"*/*/*,,*/quote", "rsc.io/quote", true},
